requirements.txt
google-generativeai
googlesearch-python
python-dotenv
